src/config.py
import os
from dotenv import load_dotenv

# Load `.env` file from the current directory
load_dotenv()

# Map the Pinecone variables
PINECONE_API_KEY = os.getenv("PINECONE_API_KEY")
PINECONE_ENV = os.getenv("PINECONE_ENV")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
GROQ_API_KEY = os.getenv("GROQ_API_KEY")
